refactor(exec-delegation): extract pure choice-safety scoring to shared
Move scoreChoiceSafety / pickSafestChoice / getChoiceCreativeCapitalCost +
gamble-key detection out of client/src/services/executiveAutoSelect.ts into
shared/utils/executiveAutoSelect.ts so the engine's autonomous-resolution path and
the client AUTO button score choices through ONE source of truth (the loyal band's
"AUTO-safe" pick is, by definition, exactly the client AUTO pick). The client
service re-exports the moved names; callers/tests unchanged.
