+++ client/src/services/executiveAutoSelect.ts
@@ -12,147 +12,19 @@
  * 4. Select top-scored executives to fill available focus slots
  */
 
-import type { Executive, RoleMeeting, DialogueChoice, ChoiceEffect } from '@shared/types/gameTypes';
+import type { Executive, RoleMeeting } from '@shared/types/gameTypes';
+import {
+  scoreChoiceSafety,
+  pickSafestChoice,
+  getChoiceCreativeCapitalCost,
+} from '@shared/utils/executiveAutoSelect';
 
-/**
- * Exec-meetings-revival PR-6 (C4) — gamble effect keys AUTO must never pick over
- * a gamble-free alternative in the same meeting. Kept local (not imported from
- * shared/engine) because the client bundle shouldn't pull in engine code just to
- * know which two keys represent risk from the player's point of view.
- */
-const GAMBLE_EFFECT_KEYS = new Set(['variance_up', 'rep_swing']);
-
-/**
- * Keys that count as "guaranteed" positive/negative value for the safety score.
- * Costs (money/creative_capital negative) get a small penalty; guaranteed
- * upside on these axes gets a small bonus. Gamble keys are handled separately.
- *
- * Exec-meetings-revival PR-7 (C5): award_chances is a guaranteed-positive
- * accumulating pool (no expiry, no roll at authoring time) — the opposite of a
- * gamble — so it belongs here, not in GAMBLE_EFFECT_KEYS. This keeps AUTO from
- * treating a real, banked prestige gain as a risk to avoid.
- */
-const VALUE_EFFECT_KEYS = ['money', 'reputation', 'creative_capital', 'award_chances'] as const;
-
-/**
- * Score a single choice's LIVE effects for risk-averseness. Higher = safer/better.
- *
- * - Any presence of a gamble key (variance_up / rep_swing) is a large penalty,
- *   in EITHER effects_immediate or effects_delayed — AUTO should never gamble on
- *   the player's behalf when a gamble-free choice exists in the same meeting.
- * - Guaranteed net-positive money/reputation/creative_capital nudges the score up;
- *   costs (negative values) nudge it down, but only a small amount — this is a
- *   tie-breaker among gamble-free choices, not the primary signal.
- */
-function scoreChoiceSafety(choice: DialogueChoice): number {
-  let score = 0;
-
-  const effectSources: (ChoiceEffect | undefined)[] = [choice.effects_immediate, choice.effects_delayed];
-
-  for (const effects of effectSources) {
-    if (!effects) continue;
-    for (const [key, value] of Object.entries(effects)) {
-      if (typeof value !== 'number') continue;
-
-      if (GAMBLE_EFFECT_KEYS.has(key)) {
-        // A gamble anywhere in the choice is a strong penalty — proportional to
-        // the magnitude so a bigger gamble is avoided even more readily.
-        score -= 100 + Math.abs(value) * 10;
-        continue;
-      }
-
-      if ((VALUE_EFFECT_KEYS as readonly string[]).includes(key)) {
-        // Small guaranteed-value signal: reward net positives, lightly penalize costs.
-        score += value > 0 ? Math.min(value, 5) : Math.max(value, -5) * 0.5;
-      }
-    }
-  }
-
-  return score;
-}
-
-/**
- * Pick the safest choice in a meeting: highest safety score wins; ties resolve
- * to the first matching choice (stable, deterministic).
- *
- * @param choices    candidate choices for the meeting
- * @param budgetLeft optional remaining Creative Capital budget. When provided,
- *   choices whose CC cost would overdraw the budget are excluded from
- *   consideration; if EVERY choice would overdraw, the least-cost choice is
- *   returned so AUTO drives CC as little negative as possible (rather than
- *   silently picking the safest-but-unaffordable option). Playtest bug #11.
- */
-export function pickSafestChoice(
-  choices: DialogueChoice[],
-  budgetLeft?: number,
-): DialogueChoice | undefined {
-  if (choices.length === 0) return undefined;
-
-  // Budget-aware path: prefer affordable choices; only if none are affordable
-  // fall back to the cheapest choice (least overdraw).
-  if (typeof budgetLeft === 'number') {
-    const affordable = choices.filter((c) => getChoiceCreativeCapitalCost(c) <= budgetLeft);
-    if (affordable.length > 0) {
-      return pickHighestSafety(affordable);
-    }
-    // No affordable choice — return the least-cost one (smallest CC spend).
-    let leastCost = choices[0];
-    let leastCostValue = getChoiceCreativeCapitalCost(leastCost);
-    for (let i = 1; i < choices.length; i++) {
-      const cost = getChoiceCreativeCapitalCost(choices[i]);
-      if (cost < leastCostValue) {
-        leastCost = choices[i];
-        leastCostValue = cost;
-      }
-    }
-    return leastCost;
-  }
-
-  return pickHighestSafety(choices);
-}
-
-/**
- * Highest safety score wins; ties resolve to the first matching choice.
- */
-function pickHighestSafety(choices: DialogueChoice[]): DialogueChoice | undefined {
-  if (choices.length === 0) return undefined;
-  if (choices.length === 1) return choices[0];
-
-  let best = choices[0];
-  let bestScore = scoreChoiceSafety(best);
-
-  for (let i = 1; i < choices.length; i++) {
-    const candidate = choices[i];
-    const candidateScore = scoreChoiceSafety(candidate);
-    if (candidateScore > bestScore) {
-      best = candidate;
-      bestScore = candidateScore;
-    }
-  }
-
-  return best;
-}
-
-/**
- * Creative Capital cost of a choice: the total CC it SPENDS, as a non-negative
- * number (0 = free or CC-positive). Reads `creative_capital` from BOTH
- * effects_immediate and effects_delayed (the engine spends on both), summing
- * only the negative components — a choice that grants CC has zero spend.
- *
- * Playtest bug #11: AUTO must never assemble a set of choices whose combined CC
- * spend exceeds the player's remaining creativeCapital.
- */
-export function getChoiceCreativeCapitalCost(choice: DialogueChoice): number {
-  let net = 0;
-  const sources: (ChoiceEffect | undefined)[] = [choice.effects_immediate, choice.effects_delayed];
-  for (const effects of sources) {
-    if (!effects) continue;
-    const value = (effects as Record<string, unknown>).creative_capital;
-    if (typeof value === 'number') net += value;
-  }
-  // Spend is the magnitude of a net negative; a net-positive/zero choice costs 0.
-  return net < 0 ? -net : 0;
-}
+// Executive Delegation arc (Tier 1): the pure safety-scoring core moved to
+// shared/utils/executiveAutoSelect so the engine's autonomous-resolution path
+// scores choices through the SAME functions (single source of truth — the loyal
+// band's "AUTO-safe" pick is exactly what this button would pick). Re-exported
+// here so existing client callers/tests keep importing them from this module.
+export { scoreChoiceSafety, pickSafestChoice, getChoiceCreativeCapitalCost };
 
 export interface AutoSelectOption {
   executive: Executive;
